Move ErrorResponse and related code/tests to their own files.
